AGRISMART/requirements.txt
Flask
Flask-WTF
WTForms
gunicorn
torch
torchvision
pillow
requests
africastalking
